refactor: Update TopBar to use drawable resource for leading icon
This commit refactors the TopBar composable to use a drawable resource ID for the leading icon instead of an ImageVector. This allows for more flexibility in the types of icons that can be used, including custom drawables.

Key changes:

Updated the leadingIcon parameter to accept @DrawableRes Int instead of ImageVector?

Update the navigationIcon composable to handle custom drawables using painterResource

Updated usage of TopBar to pass resource IDs for icons