feat: Implement FullscreenTextDialog and Integrate Translate Action
This commit introduces the FullscreenTextDialog for a better reading experience, the following changes are made:

Added FullscreenTextDialog to handle fullscreen text viewing.

Modified RecognizedTextSection to display a dialog on fullscreen click.

Passed copy, share and speak actions to the dialog.

Integrated translate action into the TextAction enum and OcrBottomSheet.

Updated the TextActionBar to call onTranslate callback.