Implement EmptyStateCard, OnboardingOverlay, and LoadingShimmer composables.
The EmptyStateCard informs the user that they have completed all available cards and offers a "Load More Cards" button.

The OnboardingOverlay provides instructions on how to interact with the flashcards (tap to flip, swipe left/right) and can be dismissed.

The LoadingShimmer displays a shimmering effect while content is loading.